% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_match.R
\name{cc_match}
\alias{cc_match}
\title{Simpel matching}
\usage{
cc_match(
  data,
  case_indicator,
  id,
  matching_factors,
  no_controls,
  replace = FALSE,
  seed = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} including both cases and potential controls.}

\item{case_indicator}{Name of the variable that indicates if an observation
is a case or a controls, codes as 1 if case and 0 if control.}

\item{id}{Name of the variable that identifies unique observations.}

\item{matching_factors}{A list of matching factors and their accompanying
matching criteria either supplied as a function or as 'exact', if
exact matching is requested for a specific variable. The matching
function should take two arguments as input where the first argument
is the value of the case and the second argument is the value of the
control. The function \code{function(x,y) abs(x - y) < 1} can for example
be used to match cases to compactors with a maximal age difference of
one year. The function \code{function(x,y) y >= x} can be used for concurrent
matching, i.e., requiering the control to be alive at the time of matching.

Example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(sex  = 'exact',
     age  = function(x,y) abs(x - y) < 1,
     time = function(x,y) y >= x)
}\if{html}{\out{</div>}}}

\item{no_controls}{The number of controls that should be matched to each
case.}

\item{replace}{Logical indicator if sampling of controls should be done with
or without replacement. Default: \code{FALSE}.}

\item{seed}{Optional seed used for the matching. Usefull for reproducibility.}

\item{verbose}{Logical indicator if default checks should be printed.
Default: \code{TRUE}.}
}
\value{
A dataset of matched cases and controls.
\describe{
\item{\code{id}}{Individuals unique id}
\item{\code{case}}{Case indicator}
\item{\code{riskset}}{Riskset number/identifier}
}
All the variables included in the supplied dataset are also returned.
}
\description{
Flexible and fast matching functions.
}
\details{
The function returns a warning if the number of available controls is
smaller than the number of controls requested for each case. In this
case all availabe controls are matched to the case.

If there are no matched availabe for a particular case, the case will
be removed from the output dataset and a warning will be shown indicating
the \code{id} of the case that has been removed.

If you want to match exact on a variable, please use the 'exact' indicator
as shown above. This will improve the computation time.
}
\examples{
require(rstpm2)

cc_match(brcancer,
         case_indicator = "hormon",
         id = "id",
         matching_factors = list(x2 = "exact",
                                 rectime = function(x, y) y >= x),
         no_controls = 1)

}
