% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpm_get_best_fit.R
\name{fpm_get_best_fit}
\alias{fpm_get_best_fit}
\title{Get best fitting FPM}
\usage{
fpm_get_best_fit(x)
}
\arguments{
\item{x}{A \code{data.frame} with AIC and BIC criteria for different FPMs obtained from
\code{fpm_test_dfs} or a list thereof.}
}
\value{
A \code{data.frame} where the first row corresponds to the model with the
lowest AIC and the second row to the model with the lowest BIC.
}
\description{
Search for the model with the lowest AIC and BIC given the output of a
\code{fpm_test_dfs} call.
}
\examples{
require(rstpm2)

data(brcancer)

test_model <- fpm_test_dfs(Surv(rectime, censrec) ~ hormon + x1,
                 dfs_bh  = 1:5,
                 dfs_tvc = list(hormon = 1:3,
                                x1     = 1:5),
                 same_dfs_tvc = TRUE,
                 data = rstpm2::brcancer)

fpm_get_best_fit(test_model)
}
