% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching}
\alias{matching}
\title{Simpel matching}
\usage{
matching(
  data,
  case_indicator,
  id,
  matching_factors,
  no_controls,
  replace = TRUE,
  seed,
  verbose = FALSE
)
}
\description{
Flexible and fast matching functions.
}
\examples{
require(rstpm2)

matching(brcancer,
         case_indicator = "hormon",
         id = "id",
         matching_factors = list(x2 = "exact",
                                 rectime = function(x, y) y >= x),
         no_controls = 1,
         verbose = TRUE)

}
