% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjust_graph.R
\name{p_adjust_graph}
\alias{p_adjust_graph}
\title{Graphical Approach to Multiple Test Procedures}
\usage{
p_adjust_graph(p, alpha, G)
}
\arguments{
\item{p}{A vector of length \emph{m} including m P-values associated with
Hypotheses \emph{j = 1,...,m} to be tested.}

\item{alpha}{Either a vector of length \emph{m} with the alpha levels against
which hypothesis j should be tested. If a hypothesis should not be
tested initially its alpha value should be \emph{0}. If only one value
is supplied for alpha it is assmued that the alpha value is splitt
equally across the hypotheses.}

\item{G}{The transition matrix associated with the graphical model assumed
for the sequential testing procedure to be used.}
}
\value{
A data.frame with testing results and adjusted P-values for each
hypotheses. Please not that results are only provided up until the
first hypothesis that failes.
}
\description{
Implementation of Bretz's graphical models for sequentially rejective
test procedures. Based on <doi: 10.1002/sim.3495>.
}
\examples{

# Example from Figure 3 in the original article
p_adjust_graph(p     = c(0.02, 0.055, 0.012),
               alpha = 0.05,
               G     = matrix(c(0  , 1/2, 1/2,
                                1/2, 0  , 1/2,
                                1/2, 1/2, 0  ),
                              nrow = 3,
                              byrow = TRUE))

# Example from Figure 5 (top left) in the original article
p_adjust_graph(p     = c(0.012, 0.02, 0.055),
               alpha = 0.05,
               G = matrix(c(0  , 1/2, 1/2,
                            0  , 0  , 1  ,
                            0  , 0  , 0  ),
                          nrow = 3,
                          byrow = TRUE))

# Example of a strict gatekeeping strategy
p_adjust_graph(p     = c(0.012, 0.055),
               alpha = c(0.05, 0),
               G = matrix(c(0  , 1,
                            0  , 0),
                          nrow = 2,
                          byrow = TRUE))
}
