% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Summary Table}
\usage{
summary_table(
  data,
  vars,
  strata = NULL,
  overall = FALSE,
  get_rates = FALSE,
  event = NULL,
  st = NULL,
  control = list(ctype = "inc.rate", method = "exact")
)
}
\arguments{
\item{data}{A \code{data.frame} that includes the data that should be summarised.}

\item{vars}{A character vector specifying the variables that should be summarised.
\code{summary_table} can only summarise variables of class \code{numeric} or
\code{factor}.}

\item{strata}{An optional character vector specifying the variables on which the
summary table should be stratified on.}

\item{overall}{If \code{TRUE} an additionall summary table will be created including the
complete dataset. This can be used in combination with \code{strata}.}

\item{get_rates}{If \code{TRUE} the summary tables includes events rates for events specified
in the \code{event} argument for each level of variables in \code{var}. Rates can
only be estimated for factor variables in \code{var}.}

\item{event}{Name of the event indicator that should be used to compute rates.
Needs to be specified if \code{get_rates == TRUE}.}

\item{st}{Name of the varaible that includes the survival time that should be used
to comptue rates. Needs to be specified if \code{get_rates == TRUE}.}

\item{control}{A \code{list} of control parameters passed to the \code{epiR::epi.conf()} call when
estimating confidence intervals for rates. This can be used to change the
type and method used for estimating confidence intervals. Check out the
documentation of \code{epiR::epi.conf()} for more information.}
}
\value{
A \code{data.table} with the following columns:
\itemize{
\item \code{varname}: Including the name of the variable or the name of the category for factor variables
\item \code{mean}: The mean of the variable, i.e. the proportion for binary variables
\item \code{p25}: The 25th percentile (only for metric variables.)
\item \code{p75}: The 75th percentile (only for metric varaibles.)
\item \code{min}: Minimum of the variable
\item \code{max}: Maximum of the variable
\item \code{n_category}: The number of non-NA observations for each variable or cateogry
\item \code{n_total}: The number of non-NA observations for the variable
\item \code{n_NA}: The number of NA observations for this variable or category
}

If strata variables were supplied to \code{summary_table} the output
\code{data.table} will include one row for each strata variable with the
corresponding value for the specific strata the row belongs to.
}
\description{
Creates stratified summary table with rates.
}
\examples{
require(rstpm2)
data("brcancer")

# Simple example
summary_table(brcancer,
              vars = c("x1", "x2", "x3"))

# Example with stratas
summary_table(brcancer,
              vars = c("x1", "x2", "x3"),
              strata = c("hormon", "x4"))

# Example with strata and rates
brcancer$x4 <- as.factor(brcancer$x4)

summary_table(brcancer,
              vars      = c("x1", "x2", "x3", "x4"),
              strata    = c("hormon"),
              get_rates = TRUE,
              event     = "censrec",
              st        = "rectime")

}
