% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table_to_workbook.R
\name{summary_table_to_workbook}
\alias{summary_table_to_workbook}
\title{Summary Table to Workbook}
\usage{
summary_table_to_workbook(workbook, summary_tables, overwrite = FALSE)
}
\arguments{
\item{workbook}{Path to an \code{.xlsx} workbook in which the summary tables should be saved
in.}

\item{summary_tables}{A list of summary tables created with \code{summary_table}.}

\item{overwrite}{This function will change the excel workbook specified in \code{workbook}.
In order to avoid unintended loss of data, the function will list the
workbook sheets that will be overwritten and asks you whether you would
like to continue. You can set \code{overwrite} to \code{TRUE} in order to avoid
the prompt. This can be useful if the function is used outside of an
interactive mode.}
}
\value{
Write the data stored in the summary tables in the specified
excel workbook.
}
\description{
Saves a list of summary tables in separate sheets of an excel workbook. This
function can be handy when creating tables for publication in combination
with some additional formatting in Excel.
}
