% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_workbook.R
\name{write_to_workbook}
\alias{write_to_workbook}
\title{Save Datasets in Excel Workbook}
\usage{
write_to_workbook(x, workbook, overwrite = FALSE)
}
\arguments{
\item{x}{A named list of datasets to be stored in a excel workbook. The list
should be of the structure \verb{list(<sheet name> = data)}. The
dataset will be stored in the excel book work sheet with the name
\verb{<sheet name>}.}

\item{workbook}{Path to an \code{.xlsx} workbook in which the datasets should be saved in.}

\item{overwrite}{This function will change the excel workbook specified in \code{workbook}.
In order to avoid unintended loss of data, the function will list the
workbook sheets that will be overwritten and asks you whether you would
like to continue. You can set \code{overwrite} to \code{TRUE} in order to avoid
the prompt. This can be useful if the function is used outside of an
interactive mode.}
}
\value{
None. Writes the data specified excel workbook sheets.
}
\description{
Saves a list of datasets in separate sheets of an excel workbook.
}
