% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_match.R
\name{cc_match}
\alias{cc_match}
\title{Simpel matching}
\usage{
cc_match(
  cases,
  controls,
  by,
  id_name,
  no_controls,
  replace = TRUE,
  seed = NULL,
  return_case_values = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{cases}{A \code{data.frame} including cases.}

\item{controls}{A \code{data.frame} including potential controls.}

\item{by}{A list of matching factors and their accompanying
matching criteria either supplied as a function or as 'exact', if
exact matching is requested for a specific variable. The matching
function should take two arguments as input where the first argument
is the value of the case and the second argument is the value of the
control. The function \code{function(x,y) abs(x - y) < 1} can for example
be used to match cases to compactors with a maximal age difference of
one year. The function \code{function(x,y) y >= x} can be used for concurrent
matching, i.e., requiering the control to be alive at the time of matching.

Example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(sex  = 'exact',
     age  = function(x,y) abs(x - y) < 1,
     time = function(x,y) y >= x)
}\if{html}{\out{</div>}}}

\item{id_name}{Name of the variable that identifies unique observations.
Please make sure that your dataset does not include a varaible named
\code{id_name}.}

\item{no_controls}{The number of controls that should be matched to each
case.}

\item{replace}{Logical indicator if sampling of controls should be done with
or without replacement. Default: \code{FALSE}.}

\item{seed}{Optional seed used for the matching. Useful for reproducibility.}

\item{return_case_values}{If \code{TRUE} the case value of the matching variable
will be returned in the output data set, e.g., if you match on time,
the controls in the output dataset will have a variable \code{case_time} which
has the value of the case's time variable. This can be useful if you
would like to start follow-up at the time of matching in your later
analysis.}

\item{verbose}{Logical indicator if default checks should be printed.
Default: \code{TRUE}.}
}
\value{
A dataset of matched cases and controls.
\describe{
\item{\code{id}}{Individuals unique id}
\item{\code{case}}{Case indicator}
\item{\code{riskset}}{Riskset number/identifier}
}
}
\description{
Flexible and fast matching functions.
}
\details{
The function returns a warning if the number of available controls is
smaller than the number of controls requested for each case. In this
case all available controls are matched to the case.

If there are no matched available for a particular case, the case will
be removed from the output dataset and a warning will be shown indicating
the \code{id} of the case that has been removed.

If you want to match exact on a variable, please use the 'exact' indicator
as shown above. This will improve the computation time.
}
\examples{

require(rstpm2)

risksets <- cc_match(cases    = brcancer[brcancer$hormon == 1, ],
                     controls = brcancer[brcancer$hormon == 0, ],
                     id = "id",
                     by = list(x2 = "exact",
                               rectime = function(x, y) y >= x),
                     no_controls = 1,
                     replace = TRUE)

head(risksets)

}
