% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpm_test_dfs.R
\name{fpm_test_dfs}
\alias{fpm_test_dfs}
\title{Tests DFs for FPMs}
\usage{
fpm_test_dfs(
  formula,
  dfs_bh = 0,
  dfs_tvc = NULL,
  by_vars = NULL,
  same_dfs_tvc = FALSE,
  data
)
}
\arguments{
\item{formula}{A R-formula that is passed to the stpm2 call.}

\item{dfs_bh}{A vector of DFs to be tested for the baseline hazard function.}

\item{dfs_tvc}{A list of variables with a vector of DFs that should be tested for, e.g.
\code{list(case = 1:10, female = 1:3)}.}

\item{by_vars}{A character vector of factor variable names used to fit stratified FPMs.
One FPM will be fitted for each combination of factor levels specified.
This is especially useful, when testing stratified FPMs.}

\item{same_dfs_tvc}{If \code{TRUE} no combinations of DFs between the different tvc variables will
be tested. Instead only DFs from the minimum DF specified until the
largest specified DF in \code{dfs_tvc} will be tested for all variables
specified in \code{dfs_tvc} at the same time.}

\item{data}{A data set that should be used to fit the FPM.}
}
\value{
A \code{data.frame} which one row per combination of baseline hazards
and tvc dfs and the corresponding AIC and BIC.
}
\description{
Test of degrees of freedom (DFs) for Flexible Parametric Survival Models
(FPMs).
}
\examples{
require(rstpm2)
require(survival)

data(brcancer)

fpm_test_dfs(Surv(rectime, censrec) ~ hormon + x1,
   dfs_bh  = 1:5,
   dfs_tvc = list(hormon = 1:3,
                  x1     = 1:5),
   data = brcancer)

fpm_test_dfs(Surv(rectime, censrec) ~ hormon + x1,
   dfs_bh  = 1:5,
   dfs_tvc = list(hormon = 1:3,
                  x1     = 1:5),
   same_dfs_tvc = TRUE,
   data = brcancer)

fpm_test_dfs(Surv(rectime, censrec) ~ hormon + x1,
   dfs_bh  = 1:5,
   dfs_tvc = list(hormon = 1:3,
                  x1     = 1:5),
   by_vars = c('x4a', 'x4b'),
   same_dfs_tvc = TRUE,
   data = brcancer)
}
